+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - added: evmChainId values for all EVM chains
+- fixed: (cosmos) Fix max spend
 
 ## 4.61.4 (2025-10-14)